Widgets: add a bit of space below the last message, so it looks better
This also required doing the proper layout of the reactions-rect-area
in doLayout, instead of drawing it "N pixels from the bottom".

This fixed some weirdness in the code too, like the hardcoded +7 which
made no sense, and the text margins which were somewhat ignored.

Now the approach is simple: layout everything in doLayout,
and just look at the "most bottom thing" to determine the height in
sizeHint().