MessageModel: fix dataChanged() not emitted for some images.
Summary:
If attach.link() contains a space, the download URL (whose path ends up
in slotFileDownloaded) had %20 instead, so the lambda didn't match.
As a result, dataChanged() wasn't emitted, and the image would overlap
messages below it.

Test Plan:
Send message with an image with a space in the name, remove cache,
restart ruqola.

Reviewers: mlaurent, kfunk

Differential Revision: https://phabricator.kde.org/D27228