Improve code block highlighting support
Summary:
Support multiple code blocks in a single message and highlight
the chunks in-between as rich text with emoticons. Furthermore,
trimm the individual text and code blocks, as we now rely on
HTML divs to seperate the blocks visually. This will allow us
to experiment with background colors for code blocks in the
future too.

Reviewers: mlaurent, dfaure

Differential Revision: https://phabricator.kde.org/D27888