Add message cache, to load missing threads for context.
Summary:
When showing a message that is a reply to a thread, we want to show the
previous message in the thread as context. But if it's so old that it
wasn't loaded, we need to
 - load the thread replies, and if it's not in there
 - load the initial thread message
Both are async, so we connect to signals and update the view once the
message comes in.

Reviewers: mlaurent

Reviewed By: mlaurent

Differential Revision: https://phabricator.kde.org/D27881