Show raw emoticon text in tooltips
Summary:
Set the raw text as title of the image and then handle help
events by looking for the format at the event position. This
then allows us to receive the raw emoticon text back and display
it in the tooltip.

We have to also improve the replace logic in the emojimanager:
Instead of blindly replacing all occurrences, only replace the
matches and don't also replace the emojis in the img title again.

Reviewers: mlaurent, dfaure

Differential Revision: https://phabricator.kde.org/D27896