Show full href in anchor tooltip
Summary:
This is esp. useful when we encounter links of the form
`[text](https://test.com)` - I do not want to click that without
knowing where it leads me to. Without this patch, I only see
`text` and not the URL. With this patch, I see the URL in a tooltip.

Reviewers: dfaure, mlaurent

Differential Revision: https://phabricator.kde.org/D27897