Revert "Lower deprecation versions to what was tested GIT_SILENT"
Until now all compiles fine. I don't want to wait that a day we see
that there is a lot of deprecated method.
We are able to fix all compile error for the moment.

My rule is not "Waiting that we switch to qt6 for preparing to migrate
it as for other kde apps".

This reverts commit d09b4b2113d288d0dc958a82573dfe537dfe6575.