Add a spin box to control the image zoom level
Summary:
The slider is far too inaccurate for me to be usable. The test is
updated and expanded to check that we can zoom by both, spinbox and
slider. This also shows that the dialog currently sizes the pixmap
according to the label size and not vice versa, which is odd. I'll
handle that in a follow up patch.

Reviewers: mlaurent, dfaure

Differential Revision: https://phabricator.kde.org/D29496