Show a notice when replying to a thread
There was no UI hint which tells the user that the current message
will be associated with a thread. This can cause accidental replies
to a thread. Added a label which hints when such is the case and also
a button that can be used to cancel a thread reply.