Only trigger completion at the end of a word.
It's really annoying to go left/left over @d and get the
completion popup at every character, if you don't actually want
to complete it.

And it was using the whole word anyway, so same offer as when we were at
the end.
Maybe we want to have some Ctrl+Space to trigger it explicitly after
messing up like @dfauHello. But only developers would think of that
shortcut, I assume... (the idea is from QtCreator)