Fix LMB-selection not copying the text when releasing the mouse over another message
This is because QAbstractItemView wouldn't call editorEvent() in that
case, because it thinks we're starting a rectangular selection to select
multiple items. All of this proves that we really don't want to use
the mouse handling from QAbstractItemView, let's just do our own
directly.

=> renamed editorEvent in messagelistdelegate to mouseEvent and called
it directly.

This also make the "repaint after a mouse event" more consistent,
since QAIV did it for press/move/release but not dblclick (!)