Show in the GUI that we're connecting to the server
Otherwise, on startup (with keychain support), it looked like the
LoginWidget was waiting for us to type the password, when in fact it's
first connecting the websocket and then moving on with the login.
Now the GUI is disabled while the websocket is connecting, so the
user knows that something is happening.