Also replace ASCII smileys with unicode smileys
This essentially removes KTextToHTML for smiley conversion, as that
doesn't look good on hidpi screens. Generally, it was odd that we
essentially had two different kinds of smileys. Now, everything is
unified and looks gorgeous even on hidpi screens.

The trick is to accumulate all possible strings into one large
regular expression pattern and then use that to do the replacement
lookups. By handling the common case of colon-escaped smileys,
the pattern isn't too large.