Don't rich-text format anything within inline code chunks
Manually iterate over the inline code chunks and replace them with
<code> tags. This prevents us from e.g. inserting smileys within
the code chunks.

Doing so means we now potentially run `generateRichText` on substrings
such as in this example line:

    foo `bar` asdf

Here, we would now run `generateRichText` on the strings "foo "
and " asdf". Because we were passing `KTextToHtml::PreserveSpaces`,
this would yield the strings "foo&nbsp;" and "&nbsp;asdf", because
KTextHtml will think these are at the start of the line. Now, this
isn't meaningful anymore, so I'm removing this flag there.

This means that we won't preserve multi-spaces within text messages
anymore, but I believe that was bogus anyways as RC online wouldn't
do that outside of code blocks either.