Allow specifying the language for code blocks
RC online uses highlight.js which allows specifying the language
for code blocks using the common form:

```<identifier>
code
```

Support this too by checking if we find a definition for the
identifier. If so, remove it from the code block and set the
highlighting definition. Otherwise, let the identifier remain
and keep on using the default highlighter (C++).