Sync layout spacing with splitter handle width in main widget
I.e. set the spacing for the roomwidgetbase and channellistwidget
layouts, such that the horizontal spacing between the channe list
and its search line edit, as well as the message list and the message
text edit, match the vertical line between the channel list and
message list widgets.

Note that in recent versions of Breeze, the search line edit is
misaligned, but that's tracked in https://bugs.kde.org/433421.