Unbreak full-resolution image attachements
The server now gives us an attachement with two urls:

image_url: the url to the small thumbnail
title_link: the url to the full-resolution image

With this patch, we will now prefer the latter and only fallback
to the former if the title_link isn't given.

For now, this unbreaks the image view for large images. Previously
we only showed the tiny thumbnail and couldn't see the large image
anymore.

In the future, we may want to leverage the thumbnail from the server.
I.e. instead of always downloading the full resolution image, we could
also only download the small version when showing it embedded in the
message view. And then we could download the large resolution only
as needed.