Implement local message logs
This was my initial reason for preferring a Qt-client over the
official one, but we had many other things to do first :-)

Since messages can be edited afterwards, and to avoid dupes when
reconnecting and history is replayed, the local storage is a SQLITE
database. The user can however export that DB as a text file, on demand,
using the action "Display / Show Channel Log".

Maybe in a few years we'll have to ask for a start date to avoid
exporting many years, if that takes too long; we'll see :)

If you think the format of the text export is similar to konversation
logs, that's, hmm, pure coincidence :-)