Call setInputTextChanged on every keypress with non-empty text
Replace the check for "no modifier" and a list of hardcoded special
characters. This way is much more robust, as previously we would not
trigger any emoji completion when we enter a key that requires a
key modifier press too. For example, this breaks completion on ':',
which is often the very first char to start an emoji. Furthermore, we
did not get any completion for ')' and similar parentheses, which are
all common for emojis too.