Implementing drag support for message attachements
Mimick the MessageDelegateHelperText API and bring it to
MessageDelegateHelperBase. This way, we can selectively drag specific
attachments from messages that might contain multiple ones. Doing so
would not be possible from within the model. And due to the custom
drag handling in the MessageListDelegate, implementing drag support
purely in the model does not work anyways.