Set unfirom item sizes on CompletionListView
This drastically improves the perceived performance when filtering
in the large emoji completion list: When the model changes or shows
up the first time, we used to layout a lot of rows which was expensive
with the complex emoji contents.

By setting a uniform size on the contents, this gets much faster and
the completion feels much snappier again.