Don't repeat the same avatar and sender in consecutive messages
When a single person sends multiple consecutive messages, we used
to repeat the avatar and sender name for every message. Now, only
the first message gets the avatar and name shown, subsequent
messages will skip that part.

The layout will still honor the space for the avatar and name though
to ensure that the messages align visually.