Repeat the sender when thread or date of consecutive message changes
I.e. not only compare the user id, also compare the thread message id
and the date of the message. Only when all three are equal will we
skip the sender and avatar when it's repeated.

This improves the visual distinction between messages when we look
at @mention lists or when the same user writes to multiple threads
consecutively.