Try to better handle NetworkSessionFailedError
This error results in tons of MessageBoxes saying:

The specified configuration cannot be used.

all at the same time, which is really annoying. What's worse is that
they keep popping up even if you close them even though the app is
working correctly and loading your messages etc.

With this change when we recieve NetworkSessionFailed, we immediately
try to logout and then log back in thereby causing a "reset" in the app.

Note that we can't just ignore the error, because if we ignore it, half
the things work and half don't. For example, reading messages, loading
rooms works but sending messages doesn't.

See: https://bugreports.qt.io/browse/QTBUG-46542