Do not fail auto-login when authentication token is invalid
Starting ruqola after some day, I am usually seeing login failures
for all my servers. The reason is that I'm too slow to unlock my
kwallet keychain, and thus no password is yet available. If that
happens when the authentication token is invalid, then we tried to
login with an empty password which is bound to fail. This patch
prevents this from happening and instead updates the password
prompt once we read it from the keychain. Then the user can manually
click login.

While this is still a bit annoying over a fully automated login in
such a scenario, it's less of a hassle than before, where we would
see a login page with a login error and an empty password prompt.