Try auto login once both, server version and password are available
The account password may arrive in a delayed fashion, esp. after
the initial boot of a machine the user first has to unlock the
keychain. If we try to auto-login beforehand, we won't have a password
and fail.

Instead, we now wait for both, the server version and the password
to become ready before we try to login.