Turn RoomListHeadingsProxyModel into a tree model
Instead of having:
 1. RoomModel determines the sections and gives an order to each channels
 2. RoomFilterProxyModel orders the channels according to that order
 3. RoomListHeadingsProxyModel adds headers inbetween channels in a linear list

This instead does:
 1. RoomModel only determines the channels' section
 2. RoomListHeadingsProxyModel creates an item for each section and
    reparents the channels to each section, forming a tree
 3. RoomFilterProxyModel sorts the channels independently in each section,
    filters the channels according to the search string and filters empty
    sections

Even though the resulting model is indeed a tree, the view is almost
unchanged, without expand/collapse capabilities, tree decoration nor
indentation.

Replaced the RocketChatAccount::updateSortOrder method and the
RocketChatAccount::needUpdateChannelView signal with the
RocketChatAccount::ownUserPreferencesChanged signal to update the
sections and ordering when sortUnreadOnTop, sortFavoriteChannels or
roomListSortOrder change.