Delete old chat accounts from RocketChatAccountModel::setAccounts
The model owns the accounts as indicated by the dtor and clear
methods. As such, when we set the accounts we need to delete the
previously used accounts too or risk leaking them.

The code should probably be modernized to use either value types
or unique_ptr to indicate the ownership more clearly. Generally
it's pretty odd to have a model own any data - usually a model
would just be an adapter to bring an existing data structure that
lives somewhere (in this case e.g. within the AccountManager) into
the Qt model/view world.