Optimize EmoticonListViewBase
Set a uniform item size and use the font metric height as row size.
This is much faster than figuring out the height of the potentially
hundreds of entries in the emoji model one by one. Especially when
using filtering the impact was very large previously. Now, filtering
is blazingly fast.