Don't reset the focus to the search line edit twice.
This is now done in a more generic way when switching rooms in the
connection between ChannelListView::roomSelected and the widget's
roomSelected signal. Don't do it again here for the ctrl-up/down
and Enter actions. Also, don't trigger re-selection of the room
we are already in apart from in the Enter case, which is supposed
to behave like clicking on the channel by mouse (which also
re-selects). Make sure that the focus in the Enter case is not on
the search line edit before doing the re-selection, as we want the
focus to explicitely go to the message line edit and be able to start
typing there, matching the semantics of what click selection does.

(cherry picked from commit f7ac35d9802a6bfbd4a29cacf6755f3337b2f77e)