Fix crash when trying to mark two channels as read quickly
If you are on a slower network and you try to mark two channels as read
quickly using the context menu, ruqola often crashes. The reason is
simple, by the time the user clicks "mark as read" the second time
model has changed and the stored QModelIndex is no longer valid leading
to a crash. To fix, use a QPersistentModelIndex

Mark as Read, is just one example. Other actions like Quit Channel etc
are also prone to this.