Speed up MessagesModel::data date handling by using UTC
To figure out if the date differs from the previous date, we can
do the comparison directly in UTC which is much faster as we do
not have to convert from UTC to local and back again.

Similarly, we also do not need to look at the local time for
the DisplayLastSeenMessage data.