Fix dont show popup menu when we found user or channel and we continue
to write text

(for example "@foo blabla")
Bug found by David

(same when we remove all character).
The problem was that we use async API => result arrived after that we
change completion type to None. => it forced to show popup.
But by default as we define completion type to none we mustn't parse
channel list => we don't show it