Fix warning about "A login operation is already ongoing"
RocketChatAccountSettings::setPassword doesn't need to
emit passwordChanged, which trigger "auto login", in addition
to the login button triggering normal login.

This is because the signal passwordChanged was only really useful
to mean "the password has been retrieved from the keychain,
use it to auto login now". So I renamed it to passwordAvailable,
and now it seems much more logical that setPassword doesn't emit it.