Use unity build for rockchatrestapi-qt lib always
It is very fast to compile, takes ~5 seconds to build from scratch with
unity build so not a problem. The unity build is configured such that
the 32 files are combined to create one TU. So this means that you won't
compile the whole library all the time, but only the files that were
part of the unity TU i.e., it will be 32 files.

This also helps reduce the library size in debug build:

Before: 43M    ./bin/librocketchatrestapi-qt.so.2.6.47
After: 8.0M    ./bin/librocketchatrestapi-qt.so.2.6.47